requirements.txt
numpy==1.19.5
Pillow==8.3.1
matplotlib==3.3.4
Cython==0.29.24
scipy==1.5.4